requirements.txt
Pillow==11.3.0
streamlit==1.42.2
